+++ controllers/cloud.redhat.com/providers/web/resources_mbop.go
@@ -5,15 +5,22 @@ import (
 	"encoding/json"
 	"fmt"
 
+	crd "github.com/RedHatInsights/clowder/apis/cloud.redhat.com/v1alpha1"
 	"github.com/RedHatInsights/clowder/controllers/cloud.redhat.com/errors"
+	obj "github.com/RedHatInsights/clowder/controllers/cloud.redhat.com/object"
 	"github.com/RedHatInsights/clowder/controllers/cloud.redhat.com/providers"
+	provutils "github.com/RedHatInsights/clowder/controllers/cloud.redhat.com/providers/utils"
+
 	rc "github.com/RedHatInsights/rhc-osdk-utils/resourceCache"
 	"github.com/RedHatInsights/rhc-osdk-utils/utils"
+
 	apps "k8s.io/api/apps/v1"
 	core "k8s.io/api/core/v1"
 	networking "k8s.io/api/networking/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 // WebBOPDeployment is the mocked bop deployment
@@ -162,4 +169,152 @@ func makeBOPIngress(p *providers.Provider) error {
 	}
 
 	return p.Cache.Update(WebBOPIngress, netobj)
+}
+
+func makeBOP(o obj.ClowdObject, objMap providers.ObjectMap, _ bool, nodePort bool) {
+	snn := providers.GetNamespacedName(o, "keycloak")
+	nn := providers.GetNamespacedName(o, "mbop")
+
+	dd := objMap[WebBOPDeployment].(*apps.Deployment)
+	svc := objMap[WebBOPService].(*core.Service)
+
+	labels := o.GetLabels()
+	labels["env-app"] = nn.Name
+
+	labeler := utils.MakeLabeler(nn, labels, o)
+
+	labeler(dd)
+
+	replicas := int32(1)
+
+	dd.Spec.Replicas = &replicas
+	dd.Spec.Selector = &metav1.LabelSelector{MatchLabels: labels}
+
+	dd.Spec.Template.ObjectMeta.Labels = labels
+
+	env := o.(*crd.ClowdEnvironment)
+	caddyImage := provutils.GetCaddyImage(env)
+
+	annotations := map[string]string{
+		"clowder/authsidecar-image":   caddyImage,
+		"clowder/authsidecar-enabled": "true",
+		"clowder/authsidecar-port":    "8090",
+		"clowder/authsidecar-config":  "caddy-config-mbop",
+	}
+
+	utils.UpdateAnnotations(&dd.Spec.Template, annotations)
+
+	envVars := []core.EnvVar{
+		{
+			Name:  "KEYCLOAK_SERVER",
+			Value: fmt.Sprintf("http://%s-keycloak.%s.svc:8080", o.GetClowdName(), o.GetClowdNamespace()),
+		},
+		{
+			Name: "KEYCLOAK_USERNAME",
+			ValueFrom: &core.EnvVarSource{
+				SecretKeyRef: &core.SecretKeySelector{
+					LocalObjectReference: core.LocalObjectReference{
+						Name: snn.Name,
+					},
+					Key: "username",
+				},
+			},
+		},
+		{
+			Name: "KEYCLOAK_PASSWORD",
+			ValueFrom: &core.EnvVarSource{
+				SecretKeyRef: &core.SecretKeySelector{
+					LocalObjectReference: core.LocalObjectReference{
+						Name: snn.Name,
+					},
+					Key: "password",
+				},
+			},
+		},
+		{
+			Name: "KEYCLOAK_VERSION",
+			ValueFrom: &core.EnvVarSource{
+				SecretKeyRef: &core.SecretKeySelector{
+					LocalObjectReference: core.LocalObjectReference{
+						Name: snn.Name,
+					},
+					Key: "version",
+				},
+			},
+		},
+	}
+
+	port := int32(8090)
+	authPort := int32(8080)
+
+	ports := []core.ContainerPort{{
+		Name:          "service",
+		ContainerPort: port,
+		Protocol:      core.ProtocolTCP,
+	}}
+
+	probeHandler := core.ProbeHandler{
+		TCPSocket: &core.TCPSocketAction{
+			Port: intstr.IntOrString{
+				Type:   intstr.Int,
+				IntVal: port,
+			},
+		},
+	}
+
+	livenessProbe := core.Probe{
+		ProbeHandler:        probeHandler,
+		InitialDelaySeconds: 10,
+		TimeoutSeconds:      2,
+	}
+	readinessProbe := core.Probe{
+		ProbeHandler:        probeHandler,
+		InitialDelaySeconds: 20,
+		TimeoutSeconds:      2,
+	}
+
+	image := provutils.GetMockBOPImage(env)
+
+	c := core.Container{
+		Name:           nn.Name,
+		Image:          image,
+		Env:            envVars,
+		Ports:          ports,
+		LivenessProbe:  &livenessProbe,
+		ReadinessProbe: &readinessProbe,
+		Resources: core.ResourceRequirements{
+			Limits: core.ResourceList{
+				"memory": resource.MustParse("200Mi"),
+				"cpu":    resource.MustParse("100m"),
+			},
+			Requests: core.ResourceList{
+				"memory": resource.MustParse("100Mi"),
+				"cpu":    resource.MustParse("50m"),
+			},
+		},
+		TerminationMessagePath:   "/dev/termination-log",
+		TerminationMessagePolicy: core.TerminationMessageReadFile,
+		ImagePullPolicy:          core.PullIfNotPresent,
+	}
+
+	dd.Spec.Template.Spec.Containers = []core.Container{c}
+	dd.Spec.Template.SetLabels(labels)
+
+	servicePorts := []core.ServicePort{
+		{
+			Name:       "mbop",
+			Port:       port,
+			Protocol:   "TCP",
+			TargetPort: intstr.FromInt(int(port)),
+		},
+		{
+			Name:       "auth",
+			Port:       authPort,
+			Protocol:   "TCP",
+			TargetPort: intstr.FromInt(int(authPort)),
+		},
+	}
+
+	utils.MakeService(svc, nn, labels, servicePorts, o, nodePort)
+
 }
